# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: SteamShare API - Steam account marketplace with points, XP, and badges
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Registered
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/User"
                  - $ref: "#/components/schemas/RegistrationChallenge"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "409":
          description: Username or email already taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/verify-registration:
    post:
      operationId: verifyRegistration
      tags: [auth]
      summary: Activate a new account with its email verification code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeInput"
      responses:
        "200":
          description: Account activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: No pending registration or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Incorrect or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/verify-2fa:
    post:
      operationId: verifyTwoFactor
      tags: [auth]
      summary: Complete login with an email 2FA code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: No pending login or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Incorrect or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/change-password:
    put:
      operationId: requestPasswordChange
      tags: [auth]
      summary: Request a password-change confirmation code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordInput"
      responses:
        "200":
          description: Confirmation code sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PasswordChangeChallenge"
        "401":
          description: Current password is incorrect
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "503":
          description: Email delivery unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /auth/change-password/confirm:
    post:
      operationId: confirmPasswordChange
      tags: [auth]
      summary: Complete a password change with its email code
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CodeInput"
      responses:
        "200":
          description: Password updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessMessage"
        "400":
          description: No pending password change or invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Incorrect or expired code
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  # Accounts
...
components:
  schemas:
...
    ApiError:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    RegisterInput:
      type: object
      required: [username, email, password]
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 30
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
    RegistrationChallenge:
      type: object
      required:
        [requiresRegistrationTwoFactor, requiresEmailVerification, email]
      properties:
        requiresRegistrationTwoFactor:
          type: boolean
          const: true
        requiresEmailVerification:
          type: boolean
          const: true
        email:
          type: string
          format: email
    CodeInput:
      type: object
      required: [code]
      properties:
        code:
          type: string
          pattern: "^[0-9]{6}$"
...
    ChangePasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
        newPassword:
          type: string
          minLength: 6
    PasswordChangeChallenge:
      type: object
      required: [requiresPasswordChangeTwoFactor, email]
      properties:
        requiresPasswordChangeTwoFactor:
          type: boolean
          const: true
        email:
          type: string
          format: email
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    User:
      type: object
      required:
        [id, username, email, points, xp, level, isAdmin, isBanned, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        avatarUrl:
          type: ["string", "null"]
        points:
          type: integer
        xp:
          type: integer
        level:
          type: integer
        badgeName:
          type: ["string", "null"]
        isAdmin:
          type: boolean
        isBanned:
          type: boolean
        banReason:
          type: ["string", "null"]
        banExpiresAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
...